Node alignment using Rankdir.
