+++ src/tilda/generation/graphviz/HTMLGenUtil.java
@@ -1,5 +1,9 @@
 package tilda.generation.graphviz;
 
+
+class HTMLGen{
+	
+}
 public class HTMLGenUtil {
 
 }
