+++ src/tilda/generation/graphviz/OsUtils.java
@@ -1,0 +1,16 @@
+package tilda.generation.graphviz;
+
+public final class OsUtils
+{
+   private static String OS = null;
+   public static String getOsName()
+   {
+      if(OS == null) { OS = System.getProperty("os.name"); }
+      return OS;
+   }
+   public static boolean isWindows()
+   {
+      return getOsName().startsWith("Windows");
+   }
+
+}
